refactor: modularize bot and api
